EasyCab_Python/requirements.txt
kafka-python
websockets
aiohttp